admin/requirements.txt
fastapi==0.68.1
gunicorn
pydantic==1.7.4
python-decouple==3.4
SQLAlchemy==1.4.27
starlette==0.14.2
uvicorn
httpx
databases[postgresql]
psycopg2-binary